Refactor for re-use in online-editor
