+++ packages/embedded-editor/package.json
@@ -29,5 +29,8 @@
       "@babel/env",
       "@babel/react"
     ]
+  },
+  "devDependencies": {
+    "filemanager-webpack-plugin": "^2.0.5"
   }
 }
