+++ packages/embedded-editor/src/embedded/index.ts
@@ -1,11 +1,11 @@
 /*
- * Copyright 2019 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *        http://www.apache.org/licenses/LICENSE-2.0
+ *       http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,12 +14,6 @@
  * limitations under the License.
  */
 
-export interface File {
-  fileName: string;
-  getFileContents: () => Promise<string | undefined>;
-}
-
-export const EMPTY_FILE = {
-  fileName: "new-file",
-  getFileContents: () => Promise.resolve("")
-};
+export { EmbeddedEditor, EmbeddedEditorRef } from "./EmbeddedEditor";
+export { EmbeddedViewer } from "./EmbeddedViewer";
+export { EmbeddedEditorRouter } from "./EmbeddedEditorRouter";
