+++ packages/embedded-editor/src/index.ts
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *        http://www.apache.org/licenses/LICENSE-2.0
+ *       http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,14 +14,5 @@
  * limitations under the License.
  */
 
-import { Router } from "@kogito-tooling/core-api";
-import * as React from "react";
-import { EnvelopeBusOuterMessageHandlerFactory } from "../editor/EnvelopeBusOuterMessageHandlerFactory";
-
-export interface EmbeddedEditorContext {
-  router: Router;
-  envelopeBusOuterMessageHandlerFactory: EnvelopeBusOuterMessageHandlerFactory;
-  iframeTemplateRelativePath: string;
-}
-
-export const EmbeddedEditorContext = React.createContext<EmbeddedEditorContext>({} as any);
+export * from "./common";
+export * from "./embedded";
