+++ packages/embedded-editor/static/envelope/envelope.html
@@ -22,6 +22,6 @@
   <body>
     <div id="loading-screen" style="z-index:100;position:relative"></div>
     <div id="envelope-app"></div>
-    <script src="index.js"></script>
+    <script src="envelope.js"></script>
   </body>
 </html>
