+++ packages/embedded-editor/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "declaration": false
+    "outDir": "dist"
   },
   "include": ["src"]
 }
