+++ packages/online-editor/package.json
@@ -11,10 +11,7 @@
   },
   "dependencies": {
     "@kogito-tooling/core-api": "0.4.3",
-    "@kogito-tooling/kie-bc-editors": "0.4.3",
-    "@kogito-tooling/microeditor-envelope": "0.4.3",
-    "@kogito-tooling/microeditor-envelope-protocol": "0.4.3",
-    "@kogito-tooling/patternfly-base": "0.4.3",
+    "@kogito-tooling/embedded-editor": "0.4.3",
     "@octokit/rest": "16.35.0",
     "@types/react-router": "5.1.1",
     "@types/react-router-dom": "5.1.1",
