+++ packages/online-editor/src/__tests__/testing_utils.tsx
@@ -14,33 +14,21 @@
  * limitations under the License.
  */
 
+import { EMPTY_FILE_DMN } from "@kogito-tooling/embedded-editor";
 import * as React from "react";
-import { GlobalContext, GlobalContextType } from "../common/GlobalContext";
-import { Routes } from "../common/Routes";
-import { EnvelopeBusOuterMessageHandlerFactory } from "../editor/EnvelopeBusOuterMessageHandlerFactory";
-import { EMPTY_FILE } from "../common/File";
-import { OnlineEditorRouter } from "../common/OnlineEditorRouter";
-import { GwtEditorRoutes } from "@kogito-tooling/kie-bc-editors";
-import { GithubService } from "../common/GithubService";
 import { Route, Switch } from "react-router";
 import { HashRouter } from "react-router-dom";
+import { GithubService } from "../common/GithubService";
+import { GlobalContext, GlobalContextType } from "../common/GlobalContext";
+import { Routes } from "../common/Routes";
 
 export function usingTestingGlobalContext(
   children: React.ReactElement,
   ctx?: Partial<GlobalContextType>
 ) {
   const usedCtx = {
-    router: new OnlineEditorRouter(
-      new GwtEditorRoutes({
-        bpmnPath: "gwt-editors/bpmn",
-        dmnPath: "gwt-editors/dmn",
-        scesimPath: "gwt-editors/scesim"
-      })
-    ),
     routes: new Routes(),
-    envelopeBusOuterMessageHandlerFactory: new EnvelopeBusOuterMessageHandlerFactory(),
-    iframeTemplateRelativePath: "../envelope/index.html",
-    file: EMPTY_FILE,
+    file: EMPTY_FILE_DMN,
     readonly: false,
     external: false,
     senderTabId: undefined,
