+++ packages/online-editor/src/common/GlobalContext.ts
@@ -14,18 +14,13 @@
  * limitations under the License.
  */
 
+import { File } from "@kogito-tooling/embedded-editor";
 import * as React from "react";
-import { Router } from "@kogito-tooling/core-api";
-import { File } from "./File";
-import { Routes } from "./Routes";
 import { GithubService } from "./GithubService";
-import { EnvelopeBusOuterMessageHandlerFactory } from "../editor/EnvelopeBusOuterMessageHandlerFactory";
+import { Routes } from "./Routes";
 
 export interface GlobalContextType {
-  router: Router;
   routes: Routes;
-  envelopeBusOuterMessageHandlerFactory: EnvelopeBusOuterMessageHandlerFactory;
-  iframeTemplateRelativePath: string;
   file: File;
   readonly: boolean;
   external: boolean;
