--- packages/online-editor/static/envelope/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <style>
-      html,
-      body,
-      div#envelope-app {
-        margin: 0;
-        border: 0;
-        padding: 0;
-        overflow: hidden;
-        height: 100%;
-      }
-    </style>
-
-    <title></title>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-  </head>
-  <body>
-    <div id="loading-screen" style="z-index:100;position:relative"></div>
-    <div id="envelope-app"></div>
-    <script src="index.js"></script>
-  </body>
-</html>
